# vecta/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vecta Platform API
  version: 1.0.0
  description: |
    Life-as-a-Service for F-1 international students.
    
    **Authentication:** All `/api/v1/*` endpoints (except webhooks) require
    `Authorization: Bearer <VectaIDToken>` (RS256 JWT).
    
    **F-1 Compliance:** `POST /api/v1/mobility/rides/accept` and
    `POST /api/v1/mobility/driver/go-online` always return `403 F1_VISA_COMPLIANCE_VIOLATION`
    regardless of token — these routes are architectural dead ends.
  contact:
    email: api@vecta.io
  license:
    name: Proprietary
    url: https://vecta.io/terms
servers:
  - url: https://api.vecta.io
    description: Production
  - url: https://api.staging.vecta.io
    description: Staging
  - url: http://localhost:4000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    VectaIDToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT issued by Vecta identity service. Contains zero PII.
...
paths:
...
  /webhooks/didit:
    post:
      summary: Didit NFC verification webhook
      tags: [Webhooks]
      security: []
      responses:
        '200': { description: Accepted }
        '401': { description: HMAC verification failed }
...